requirements.txt
streamlit==1.29.0
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
openai==1.58.1
pandas==2.1.4
python-dotenv==1.0.0
azure-ai-documentintelligence==1.0.0b1
Pillow==10.1.0
